+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Do not check files in `vendor/`
+
 ## [0.0.3] - 2026-02-15
 
 - exit with non-zero exit code if any violations are found